+++ examples/resources/routeros_system_script/import.sh
@@ -1,0 +1,3 @@
+#The ID can be found via API or the terminal
+#The command for the terminal is -> :put [/system/script get [print show-ids]]
+terraform import routeros_system_script.script "*0"