+++ examples/resources/routeros_wifi/resource.tf
@@ -1,0 +1,6 @@
+resource "routeros_wifi" "wifi1" {
+  configuration = {
+    manager = "capsman"
+  }
+  name = "wifi1"
+}